# backend-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Receipt Tracker API",
    "description": "API for managing receipts and transactions with AI-powered data extraction using Google Gemini",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/audit/logs": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get audit logs (Admin only)",
        "description": "Query audit logs with optional filters. Requires admin privileges.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by user ID"
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by action type (e.g., LOGIN, UPLOAD_RECEIPT)"
          },
          {
            "name": "resource_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by resource type (e.g., Receipt, User)"
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter logs after this date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter logs before this date (ISO format)"
          },
          {
            "name": "success",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by success status"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 1000
            },
            "description": "Maximum number of logs to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of logs to skip (pagination)"
          },
          {
            "name": "order_by",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "description": "Sort order by timestamp"
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/audit/logs/{log_id}": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get specific audit log (Admin only)",
        "description": "Retrieve a specific audit log by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "log": {
                      "$ref": "#/components/schemas/AuditLog"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/audit/user/{user_id}/activity": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get user activity",
        "description": "Get activity history for a specific user. Users can view their own activity, admins can view any user's activity.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User activity retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/audit/user/me/activity": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get current user's activity",
        "description": "Get activity history for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User activity retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/audit/failed-logins": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get failed login attempts (Admin only)",
        "description": "Get recent failed login attempts for security monitoring",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by username"
          },
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 24
            },
            "description": "Number of hours to look back"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Failed logins retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "failed_logins": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "hours": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/audit/resource/{resource_type}/{resource_id}/history": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get resource history (Admin only)",
        "description": "Get audit history for a specific resource",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resource_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Receipt"
          },
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resource_type": {
                      "type": "string"
                    },
                    "resource_id": {
                      "type": "integer"
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/audit/statistics": {
      "get": {
        "tags": ["Audit"],
        "summary": "Get audit statistics (Admin only)",
        "description": "Get statistics about audit logs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["today", "week", "month"]
            },
            "description": "Predefined period for statistics"
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statistics": {
                      "$ref": "#/components/schemas/AuditStatistics"
                    },
                    "period": {
                      "type": "object",
                      "properties": {
                        "start_date": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "end_date": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/audit/export": {
      "get": {
        "tags": ["Audit"],
        "summary": "Export audit logs (Admin only)",
        "description": "Export audit logs as CSV or JSON",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["csv", "json"],
              "default": "csv"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1000,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export successful",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditLog"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "exported_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "error": {
            "type": "string",
            "example": "Detailed error information"
          }
        }
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1,
            "nullable": true
          },
          "username": {
            "type": "string",
            "example": "john_doe",
            "nullable": true
          },
          "user_role": {
            "type": "string",
            "example": "System Admin",
            "nullable": true
          },
          "action": {
            "type": "string",
            "example": "LOGIN"
          },
          "resource_type": {
            "type": "string",
            "example": "Receipt",
            "nullable": true
          },
          "resource_id": {
            "type": "integer",
            "example": 123,
            "nullable": true
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "endpoint": {
            "type": "string",
            "example": "/api/auth/login"
          },
          "ip_address": {
            "type": "string",
            "example": "192.168.1.100",
            "nullable": true
          },
          "user_agent": {
            "type": "string",
            "example": "Mozilla/5.0...",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-05T10:30:00Z"
          },
          "status_code": {
            "type": "integer",
            "example": 200,
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "example": true
          },
          "error_message": {
            "type": "string",
            "example": null,
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "example": {
              "duration_seconds": 0.123,
              "query_params": {}
            },
            "nullable": true
          },
          "session_id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "nullable": true
          }
        }
      },
      "AuditStatistics": {
        "type": "object",
        "properties": {
          "total_logs": {
            "type": "integer",
            "example": 1523
          },
          "successful_actions": {
            "type": "integer",
            "example": 1498
          },
          "failed_actions": {
            "type": "integer",
            "example": 25
          },
          "unique_users": {
            "type": "integer",
            "example": 15
          },
          "success_rate": {
            "type": "number",
            "format": "float",
            "example": 98.36
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}